Test if paths exist in build script.

+++ tools/BuildStatic.ps1
@@ -30,14 +30,23 @@ If (Test-Path "$($sitePath)\Content\Images")
     Copy-Item "$($sitePath)\Content\Images\*" -Destination "$($outputPath)\Content\Images" -Force -Recurse
 }
 
-New-Item "$($outputPath)\Content" -ItemType Directory
-Copy-Item "$($sitePath)\Content\*.css" -Destination "$($outputPath)\Content" -Force -Recurse
+If (!(Test-Path "$($outputPath)\Content"))
+{
+	New-Item "$($outputPath)\Content" -ItemType Directory
+	Copy-Item "$($sitePath)\Content\*.css" -Destination "$($outputPath)\Content" -Force -Recurse
+}
 
-New-Item "$($outputPath)\fonts" -ItemType Directory
-Copy-Item "$($sitePath)\fonts\*" -Destination "$($outputPath)\fonts" -Force -Recurse
+If (!(Test-Path "$($outputPath)\fonts"))
+{
+	New-Item "$($outputPath)\fonts" -ItemType Directory
+	Copy-Item "$($sitePath)\fonts\*" -Destination "$($outputPath)\fonts" -Force -Recurse
+}
 
-New-Item "$($outputPath)\Scripts" -ItemType Directory
-Copy-Item "$($sitePath)\Scripts\*.js" -Destination "$($outputPath)\Scripts" -Force -Recurse
+If (!(Test-Path "$($outputPath)\Scripts"))
+{
+	New-Item "$($outputPath)\Scripts" -ItemType Directory
+	Copy-Item "$($sitePath)\Scripts\*.js" -Destination "$($outputPath)\Scripts" -Force -Recurse
+}
 
 # Debug print output.
 Write-Host "Content of output:"
